requirements.txt
apted==1.0.3
arxiv==2.1.3
beautifulsoup4==4.13.3
click==8.1.8
datasets==3.2.0
distance==0.1.3
fastapi==0.115.8
feedparser==6.0.11
filetype==1.2.0
ftfy==6.3.1
lxml==5.3.0
markdown2==2.5.3
markdownify==0.14.1
numpy==2.2.2
pdfplumber==0.11.5
pdftext==0.5.1
Pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_settings==2.7.1
pypdfium2==4.30.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
rapidfuzz==3.12.1
regex==2024.11.6
Requests==2.32.3
scikit_learn==1.6.1
setuptools==75.8.0
starlette==0.45.3
streamlit==1.42.0
surya==1.0.0
tabulate==0.9.0
texify==0.2.1
torch==2.6.0+cu118
tqdm==4.67.1
transformers==4.48.2
uvicorn==0.34.0
